Extract get-background
