+++ packages/alfa-rules/src/common/dom/get-colors.ts
@@ -3,6 +3,6 @@
  */
 export * from "./get-colors/color";
 export * from "./get-colors/color-error";
-export * from "./get-colors/get-colors";
+export * from "./get-colors/get-background";
 export * from "./get-colors/get-foreground";
 export * from "./get-colors/get-layers";
