+++ packages/alfa-rules/src/common/dom/get-colors/get-background.ts
